# Optimize DropCountr integration performance and reduce API calls